requirements.txt
streamlit
moviepy
gtts
pillow
colorthief